+++ Modules/Intent.Modules.Common.CSharp/FactoryExtensions/CSharpTypesCache.cs
@@ -1,9 +1,13 @@
-﻿using System;
+﻿#nullable enable
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Transactions;
 using Intent.Engine;
+using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.CSharp.TypeResolvers;
+using Intent.Modules.Common.CSharp.VisualStudio;
 using Intent.Modules.Common.Plugins;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.TypeResolution;
@@ -20,13 +24,14 @@ namespace Intent.Modules.Common.CSharp.FactoryExtensions
         /// <inheritdoc />
         public override string Id => "Intent.Modules.Common.CSharp.FactoryExtensions.KnownCSharpTypesCache";
 
-        private static TypeRegistry _knownTypes;
-        private static TypeRegistry _outputTargetNames;
+        private static TypeRegistry? _knownTypes;
+        private static TypeRegistry? _outputTargetNames;
         private static readonly TypeRegistry Empty = new([]);
-        private static readonly HashSet<string> ManuallyAddedKnownTypes = new();
+        private static readonly HashSet<string> ManuallyAddedKnownTypes = [];
+        private static readonly Dictionary<string, HashSet<string>> GlobalUsingsByProjectId = [];
 
         /// <inheritdoc />
-        public override int Order { get; set; } = int.MinValue;
+        public override int Order { get; set; }
 
         /// <inheritdoc />
         protected override void OnAfterTemplateRegistrations(IApplication application)
@@ -40,22 +45,97 @@ namespace Intent.Modules.Common.CSharp.FactoryExtensions
                 .Union(commonKnownTypes)
                 .Union(ManuallyAddedKnownTypes)
                 .ToArray();
-            var outputTargetNames = application.OutputTargets.Select(x => x.Name).ToArray();
 
             _knownTypes = new TypeRegistry(knownTypes);
 
             var instanceRegistryKnownTypes = TemplateInstanceRegistry.GetRegisteredTypes().OfType<CSharpResolvedTypeInfo>();
-            foreach (var otherInstance in instanceRegistryKnownTypes )
+            foreach (var otherInstance in instanceRegistryKnownTypes)
             {
                 _knownTypes.Add(otherInstance.Namespace, otherInstance.Name);
             }
 
+            var outputTargetNames = application.OutputTargets.Select(x => x.Name).ToArray();
             _outputTargetNames = new TypeRegistry(outputTargetNames);
+
+            PopulateGlobalUsings(application);
         }
 
-        internal static void AddKnownType(string fullyQualifiedTypeName) => ManuallyAddedKnownTypes.Add(fullyQualifiedTypeName);
+        private static void PopulateGlobalUsings(IApplication application)
+        {
+            var csharpTemplates = application.FindTemplateInstances<IIntentTemplate>(TemplateDependency.OfType<IIntentTemplate>());
+            var csharpTemplatesByProject = csharpTemplates.GroupBy(GetOutputProjectId);
 
-        private static IEnumerable<string> GetCommonKnownTypes()
+            foreach (var grouping in csharpTemplatesByProject)
+            {
+                if (!GlobalUsingsByProjectId.TryGetValue(grouping.Key, out var globalUsings))
+                {
+                    globalUsings = [];
+                    GlobalUsingsByProjectId.Add(grouping.Key, globalUsings);
+                }
+
+                var templates = grouping.ToArray();
+
+                var alreadyProcessedFiles = new HashSet<string>(templates.Length);
+                var csProjFolder = default(string);
+
+                foreach (var template in templates)
+                {
+                    var path = template.FileMetadata.GetFilePath().Replace('\\', '/');
+
+                    if (template.TryGetExistingFilePath(out var existingFilePath))
+                    {
+                        alreadyProcessedFiles.Add(existingFilePath);
+                    }
+
+                    if (Path.GetExtension(path).Equals(".csproj", StringComparison.OrdinalIgnoreCase) &&
+                        File.Exists(path))
+                    {
+                        csProjFolder = Path.GetDirectoryName(path);
+                    }
+
+                    if (template is not ICanContainGlobalUsings canContainGlobalUsings)
+                    {
+                        continue;
+                    }
+
+                    globalUsings.UnionWith(canContainGlobalUsings.GetGlobalUsings());
+                }
+
+                if (csProjFolder == null)
+                {
+                    continue;
+                }
+
+                var filePaths = Directory.EnumerateFiles(csProjFolder, "*.cs", new EnumerationOptions { MatchCasing = MatchCasing.CaseInsensitive, RecurseSubdirectories = true });
+                foreach (var filePath in filePaths)
+                {
+                    if (!alreadyProcessedFiles.Add(filePath.Replace('\\', '/')))
+                    {
+                        continue;
+                    }
+
+                    globalUsings.UnionWith(File.ReadAllLines(filePath)
+                        .Select(x => x.Trim())
+                        .Where(x => x.StartsWith("global using "))
+                        .Select(x => x["global using ".Length..^1]));
+                }
+            }
+        }
+
+        private static string GetOutputProjectId(IIntentTemplate templateInstance) => templateInstance.OutputTarget.GetProject().Id;
+
+        internal static void AddKnownType(string fullyQualifiedTypeName)
+        {
+            if (_knownTypes == null)
+            {
+                ManuallyAddedKnownTypes.Add(fullyQualifiedTypeName);
+                return;
+            }
+
+            _knownTypes.Add(fullyQualifiedTypeName);
+        }
+
+        private static string[] GetCommonKnownTypes()
         {
             // This forces these assemblies to be loaded prior to the call to
             // AppDomain.CurrentDomain.GetAssemblies()
@@ -74,41 +154,15 @@ namespace Intent.Modules.Common.CSharp.FactoryExtensions
                 .Where(x => x.GetName().Name?.Split('.')[0] == "System")
                 .SelectMany(c => c.GetExportedTypes())
                 .Where(x => !x.ContainsGenericParameters)
-                .Select(x => x.FullName)
+                .Select(x => x.FullName!)
                 .OrderBy(x => x)
                 .ToArray();
         }
 
-        internal static TypeRegistry GetKnownTypes()
-        {
-            if (_knownTypes == null)
-            {
-                // TODO: Re-add this warning once we've resolved the issue of some decorators calling this during their construction: https://dev.azure.com/intentarchitect/Intent%20Architect/_workitems/edit/1282
-                //Logging.Log.Warning($"{nameof(GetKnownTypesByNamespace)} is being called before " +
-                //                    "Template Registration has been completed. Ensure that methods " +
-                //                    "like GetTypeName and UseType are not being used in template " +
-                //                    $"constructors.{Environment.NewLine}{Environment.StackTrace}");
+        internal static TypeRegistry GetKnownTypes() => _knownTypes ?? Empty;
 
-                return Empty;
-            }
+        internal static TypeRegistry GetOutputTargetNames() => _outputTargetNames ?? Empty;
 
-            return _knownTypes;
-        }
-
-        internal static TypeRegistry GetOutputTargetNames()
-        {
-            if (_knownTypes == null)
-            {
-                // TODO: Re-add this warning once we've resolved the issue of some decorators calling this during their construction: https://dev.azure.com/intentarchitect/Intent%20Architect/_workitems/edit/1282
-                //Logging.Log.Warning($"{nameof(GetKnownTypesByNamespace)} is being called before " +
-                //                    "Template Registration has been completed. Ensure that methods " +
-                //                    "like GetTypeName and UseType are not being used in template " +
-                //                    $"constructors.{Environment.NewLine}{Environment.StackTrace}");
-
-                return Empty;
-            }
-
-            return _outputTargetNames;
-        }
+        internal static IReadOnlySet<string> GetGlobalUsings(ICSharpTemplate template) => GlobalUsingsByProjectId.TryGetValue(template.Project.Id, out var globalUsings) ? globalUsings : [];
     }
 }